Various fixes
The CREATE EXTENSION call needs to be here because it needs a superuser
to run it and this is the only time we run with one.

Then create the migrations user first. Then switch roles to the
migrations user to grant permissions to the consignment api user so it
can use the tables when the database migrations are run.